requirements-dev.txt
PyQt6-tools